fix tests

+++ tests/logging-output-tests.js
@@ -23,6 +23,7 @@ describe('logging', function() {
     }
   };
   var logging;
+  //  = require('../lib/logging');
 
   var log;
   beforeEach(function(done) {
@@ -32,15 +33,16 @@ describe('logging', function() {
     done();
   });
 
-  afterEach(function() {
+  afterEach(function(done) {
     //List all files in the test logs and get rid of em.
     const logDirPath = './tests/log';
     var files = fs.readdirSync('./tests/log');
-    files.forEach((f) => {
-      fs.unlinkSync(path.join(logDirPath, f));
+    files.forEach(f => {
+      if (fs.existsSync(path.join(logDirPath, f))) fs.unlinkSync(path.join(logDirPath, f));
     });
+    done();
   });
-  
+
   it('exposes a _calculateLogFilePath function', () => {
     assert.ok(logging._calculateLogFilePath);
   });
@@ -57,15 +59,17 @@ describe('logging', function() {
     assert.equal(content.toString(), 'foo');
   });
 
-  it('will log multiple messages to file', () => {
+  it('will log multiple messages to file', done => {
     var whereToFindLog = logging._calculateLogFilePath('./tests/log', '1', 1);
     log.writeLogRecord({ level: 'info', msg: 'foo1' });
     log.writeLogRecord({ level: 'info', msg: 'foo2' });
     //A small work around for ensuring that write buffers are cleared to files before checking if data is present in logs.
-    process.nextTick(() => {
-      var content = fs.readFileSync(whereToFindLog);
-      assert.equal(content.toString(), 'foo1foo2');
-    });
+    setTimeout(() => {
+      console.log('whereToFindLog', whereToFindLog);
+      var content = fs.readFileSync(whereToFindLog, 'utf8');
+      console.log('content', content);
+      assert.deepStrictEqual(content, 'foo1foo2');
+      done();
+    }, 100);
   });
-
 });
